### 2024-01-12 #6218 @editorialbot
[PRE REVIEW]: ThermoParser: Streamlined Analysis of the Thermoelectric Properties
**Submitting author:** @kbspooner (<a href="http://orcid.org/0000-0002-4078-3148">Kieran B. Spooner</a>)
**Repository:** https://github.com/SMTG-Bham/ThermoParser
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** 3.0.0
