### 2024-01-12 #6218 @Kevin-Mattheus-Moerman
Note to other editors, this is an expanded/improved re-submission of: https://github.com/openjournals/joss-reviews/issues/3012
