### 2024-01-31 #6218 @mkhorton
This looks like an excellent tool, I’d definitely have thoughts! Compliments to @kbspooner and team.

Unfortunately I have to decline the review due to time constraints. I was quite delayed with the last JOSS submission I reviewed and I’d prefer to defer to other people who would be able to review more promptly out of consideration to the authors.
