### 2024-02-01 #6218 @jranalli
I'd be happy to do a review for JOSS at some point, but I'm sorry to say that this particular topic is well outside my area of expertise. I don't have any experience related to material science or with the types of data that this is designed to handle that would allow me to provide any real insight into this software. So, I don't think I'd be comfortable trying to review it. Sorry!
