### 2024-02-06 #6218 @mbarzegary
@enricgrau since the tool is focused on the visualization and data analysis (rather than the core computation) and you have experience in open source development of such tools in materials science, I believe than you can contribute to the review of this submission (and improving it if needed). Please let me know if you have time to do it.
