### 2024-02-06 #6218 @mbarzegary
Hi @ksunden, @espottesmith, @hgandhi2411, @fnattino 👋
Considering that you have already reviewed for JOSS, would any of you be willing to review this submission? As you know, we carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Paper: https://raw.githubusercontent.com/openjournals/joss-papers/joss.06218/joss.06218/10.21105.joss.06218.pdf
Software: https://github.com/SMTG-Bham/ThermoParser
