### 2024-02-12 c74129a
Update ref.bib
add missing dois
