+++ ref.bib
@@ -179,6 +179,7 @@
   year      = {2022},
   pages     = {111152},
   volume    = {206},
+  doi       = {10.1016/j.commatsci.2021.111152},
   publisher = {Elsevier},
 }
 
@@ -190,6 +191,7 @@
   number    = {12},
   pages     = {2140},
   volume    = {181},
+  doi       = {10.1016/j.cpc.2010.08.027},
   publisher = {Elsevier},
 }
 
@@ -211,6 +213,7 @@
   year      = {2018},
   pages     = {140},
   volume    = {231},
+  doi       = {10.1016/j.cpc.2018.05.010},
   publisher = {Elsevier},
 }
 
@@ -221,6 +224,7 @@
   year      = {2021},
   pages     = {107970},
   volume    = {264},
+  doi       = {10.1016/j.cpc.2021.107970},
   publisher = {Elsevier},
 }
 
@@ -232,6 +236,7 @@
   number    = {1},
   pages     = {46},
   volume    = {6},
+  doi       = {10.1038/s41524-020-0316-7},
   publisher = {Nature Publishing Group UK London},
 }
 
@@ -242,6 +247,7 @@
   year      = {2023},
   pages     = {108670},
   volume    = {287},
+  doi       = {10.1016/j.cpc.2023.108670},
   publisher = {Elsevier},
 }
 
@@ -266,6 +272,7 @@
   number    = {5},
   pages     = {1800184},
   volume    = {2},
+  doi       = {10.1002/adts.201800184},
   publisher = {Wiley Online Library},
 }
 
