### 2024-02-12 e2801db
Update ref.bib
added the remaining dois, other missing data and references
