+++ ref.bib
@@ -67,6 +67,7 @@
   journal = {Computer Physics Communications},
   title   = {{ShengBTE:} a Solver of the {B}oltzmann Transport Equation for Phonons},
   year    = {2014},
+  number  = {6},
   pages   = {1747},
   volume  = {185},
   doi     = {10.1016/j.cpc.2014.02.015},
@@ -169,6 +170,10 @@
   journal   = {Journal Materials Chemistry C},
   title     = {{Sr}$_{2}${Sb}$_{2}${O}$_{7}$: a Novel Earth Abundant Oxide Thermoelectric},
   year      = {2023},
+  number    = {27},
+  pages     = {9124},
+  volume    = {11},
+  doi       = {10.1039/D3TC01003A},
   publisher = {Royal Society of Chemistry},
 }
 
@@ -203,6 +208,7 @@
   number    = {20},
   pages     = {1800246},
   volume    = {8},
+  doi       = {10.1002/aenm.201800246},
   publisher = {Wiley Online Library},
 }
 
@@ -284,6 +290,7 @@
   number    = {32},
   pages     = {16813},
   volume    = {10},
+  doi       = {10.1039/D2TA04160J},
   publisher = {Royal Society of Chemistry},
 }
 
@@ -295,6 +302,7 @@
   number    = {18},
   pages     = {7441},
   volume    = {33},
+  doi       = {10.1021/acs.chemmater.1c02164},
   publisher = {ACS Publications},
 }
 
@@ -306,7 +314,31 @@
   number    = {10},
   pages     = {2709},
   volume    = {8},
+  doi       = {10.1039/D1MH00764E},
   publisher = {Royal Society of Chemistry},
+}
+
+@article{Han2024,
+  title     = {Discovery of Multi-Anion Antiperovskites {X}$_{6}${NFSn}$_{2}$ ({X}= {Ca}, {Sr}) as Promising Thermoelectric Materials by Computational Screening},
+  author    = {Han, Dan and Zhu, Bonan and Cai, Zenghua and Spooner, Kieran B and Rudel, Stefan S and Schnick, Wolfgang and Bein, Thomas and Scanlon, David O and Ebert, Hubert},
+  journal   = {Matter},
+  volume    = {7},
+  number    = {1},
+  pages     = {158},
+  year      = {2024},
+  doi       = {10.1016/j.matt.2023.10.022},
+  publisher = {Elsevier},
+}
+
+@article{Willis2023,
+  title     = {On the Possibility of p-Type Doping in Barium Stannate},
+  author    = {Willis, Joe and Spooner, Kieran B and Scanlon, David O},
+  journal   = {Applied Physics Letters},
+  volume    = {123},
+  number    = {16},
+  year      = {2023},
+  doi       = {10.1063/5.0170552},
+  publisher = {AIP Publishing},
 }
 
 @Comment{jabref-meta: databaseType:bibtex;}
