### 2024-02-12 92e4e5e
Update paper.md
updated references
