### 2024-02-12 #6340 @kbspooner
phono3py-power-tools doesn't have a doi but a url is included.
