### 2024-02-12 #6340 @espottesmith
I've gone through the paper, the documentation, and the code/examples. As a supplement to the checklist that I've filled out above, here are my (scattered) thoughts about this work as it exists now.

Though I've been in research environments where folks focused on thermoelectrics, I don't myself perform thermoelectric research. As a result, I'd say I have a pretty mid-level understanding of these materials and how they're analyzed, and so I will trust the other reviewers to comment on the more technical aspects of the code (for instance, are all of the numerical transformations fully accurate?). Thankfully, this package is mainly focused on data processing and visualization. I don't think that any lack of familiarity with the science adversely affects my ability to test and evaluate ThermoParser.

- Install went mostly smoothly, but I did for some reason need to install matplotlib separately before "pip install ." would work. Not a huge issue, in my opinion.
- All tests pass, and testing was incredibly smooth (3 seconds for 150 tests - impressive!)
- I thought that the tutorials were helpful to understand how this code works, and very much appreciated
- I did not run all examples, but I did try several, using both the command-line and Python scripts. I was not actually able to run all of the tests that I tried. For instance, in 04-layered-plots, I got the error "Exception: File not found, please use get-data.sh". There are actually several get-data.sh scripts (
	./cumkappa/get-data.sh
	./waterfall/get-data.sh
	./heatmaps/get-data.sh
	./projected-phonons/get-data.sh
)
	but none in the directory that I was testing. It was not immediately clear which of these scripts I should run. I'd appreciate if documentation for obtaining data were improved!
- On a related note, I tried running a get-data.sh script (in ThermoParser/examples/waterfall) on my personal computer (using EndeavourOS Linux x86_64; happy to provide more system details if relevant and needed), using "bash ./get-data.sh". I got the errors:
	./get-data.sh: line 2: $'\r': command not found
	./get-data.sh: line 5: $'\r': command not found
	./get-data.sh: line 11: syntax error: unexpected end of file
which seems to indicate that these scripts are formatted for DOS-based systems. A little strange, given that they're using Bash. I was able to get around this by using a simple conversion ("sed -i 's/\r$//' get-data.sh"). However, this manipulation ideally would not be necessary. get-data.sh and similar scripts should be able to be run on Unix-based systems without modification.
- Love the thoughtful consideration of defaults and configuration. I've had to hack some science plotting tools before that shall remain nameless, and I would have loved to have just edited a *.yaml file instead.
- Some of the links in the documentation are broken: for instance, see the link for "tutorials" here: https://smtg-bham.github.io/ThermoParser/#command-line-interface-cli
- I think the code is quite well written. There are some inconsistencies - not all functions follow the same docstring format, for instance (compare e.g. tp/cli/options.py and tp/setup/vasp.py) - but I could generally tell what was happening in a given function.
- I'm curious why there is a utility to run BoltzTraP but none of the other supported codes. The authors note that some of the pymatgen code is slow; is this not true for other tools? If some of the pymatgen code is problematic, why did the authors not simply try to submit pull requests into pymatgen? (Disclosure: I'm a somewhat active contributor to pymatgen, though not to any of the BoltzTraP-related code)
- Being able to process experimental data sounds potentially challenging, but extremely useful. I encourage the authors to pursue this avenue in the future!

With small modifications that I've highlighted, I think this package and associated paper are suitable for publication in JOSS, and I give them my recommendation.
