### 2024-02-15 SMTG-Bham/ThermoParser#71 @enricgrau
[joss review] review from @enricgrau
Regarding [this](https://github.com/openjournals/joss-reviews/issues/6340#issuecomment-1939521565) JOSS review

Overall, the library shows a clear motivation and solution, along with comprehensive documentation, examples, and tutorials. I’d support for publication in JOSS after addressing my comments and concerns herein.

**Manuscript**
The paper is well written and structured. It presents a clear motivation and solution with the library. The authors do a good job of describing the library, its structure, and its use. Here are some observations to improve the current state of the manuscript:
1.	The first sentence states that “thermoelectric materials could be an important renewable energy source to help slow the climate crisis”. Why? How? Examples of this? This feels like a random phrase without further context. However, I do not think it is necessary to include it in the manuscript as the identified problem is motivation and justification enough for the library. In other words, I suggest removing this line or developing a broader context on _why_ and _how_ these materials can help with the climate crisis.
2.	The authors claim the library has been used in the literature, however, explicit mention of the library is not present in the cited articles. There is mention of the library ``ThermoPlotter`` in ``Kavenagh (2021)``, ``Herring Rodriguez (2023)``, ``Brlec (2022)``, and ``Spooner (2021)`` which looks to be a previous version of ``ThermParser``. Clarifying this in the manuscript would increase confidence in this claim. However, I did not find any mention in neither ``Spooner (2020)`` or ``Einhorn (2020)``. Even though I’m not specifically questioning the use of the library on those specific articles, I don’t think it corresponds to include them due to the lack of any explicit mention of it. Please correct me if there is any kind of mention of the library in said articles that I did not catch.
3.	I believe there is a typo at L.16: BoltzTrap(2), other mentions of the library are L.50: BoltzTraP, L.155:  BoltzTraP…
4.	Phrasing could be better in L.95-97. Is the “;” a typo? Maybe a bullet point structure could improve the understanding of future directions of the library (if there are several)
5.	For Author Contribution I recommend using [CRediT](https://credit.niso.org/) for a clearer and fair characterization of the authors’ contribution.

**Documentation**
  _Gallery_
  -	The Tutorials links are broken
  -	Clicking in figures links to a broken URL
  -	The reference is not in the correct format. It displays the ``bibtex`` entry rather than “1. Maradudin, ….”

  _Tutorials_
  It is great that you show both ``CLI`` and ``Python`` options and the thorough explanation of what each stage is doing and which lines they correspond to. Good job! However, I did have trouble running them. Here are some comments:
  -	There is no context on the data. You should include what kind of data the tutorial is working with to better understand its application (e.g. Where does the data come from? How was it measured?)
  -	Links in the titles are broken (in https://smtg-bham.github.io/ThermoParser/tutorials.html), but in the menu they work.
  -	Tutorial-02: no file ``kappa-m404021.hdf5`` is incldued
  -	Tutorial-03: with CLI I get ``raise ValueError(f"{v!r} is not a valid value for {k}") ValueError: "'#f0901f'" is not a valid value for color``. Python works.
  -	Tutorial-04: ``<sub>3</sub>`` appears as text
  -	Tutorial-04: file ``kappa-m363636.hddf5`` not included
  -	Tutorial-05: file ``kappa-m363636.hddf5`` not included
  -	The main purpose of the library is for visualization. However, the authors claim that it is also for “analyzing”. After reviewing the documentation this is not clear to me. It would be helpful to strengthen this claim by including an example where this is the main focus. I could be misunderstanding this however. Please comment on this.

_Installation_
-	Build badge shows “unknown”
-	The “gallery” link is broken
-	I tested for ``Windows`` and installation works.

_Python Package_
-	Some functions (for example ``tp.setup.vasp.gen_ibz``) show the first line highlighted in gray in the docs.

**Repository**
-	What version are we working with? There is only 1 release (``v1.0.0``), but the title says ``3.0.0`` (?). It would be great to explicitly mention the current version available and also using more clear title names for future releases (``v1.0.0`` instead of “Launch day”).

**Tests**
Tests are good. They run smoothly with 98% and 89% coverage for ``calculate.py`` and ``settings.py``, respectively. Good job.
