### 2024-02-15 #6340 @enricgrau
@mbarzegary @kbspooner I have added my review in [this issue](https://github.com/SMTG-Bham/ThermoParser/issues/71).
