### 2024-02-16 188f2b8
added contour and contourf to heatmaps
heatmap functions now can add contours or discretise the colourmap with contour(f)
