+++ tp/plot/heatmap.py
@@ -47,8 +47,8 @@ def add_heatmap(ax, x, y, c, xinterp=None, yinterp=None, kind='linear',
                 xscale='linear', yscale='linear', cscale='linear', xmin=None,
                 xmax=None, ymin=None, ymax=None, cmin=None, cmax=None,
                 colour='viridis', undercolour=None, overcolour=None,
-                contours=None, contourcolours='black', contourkwargs=None,
-                **kwargs):
+                discrete=False, levels=None, contours=None,
+                contourcolours='black', contourkwargs=None, **kwargs):
     """Adds a heatmap to a set of axes.
 
     Formats limits, parses extra colourmap options, makes sure data
@@ -98,6 +98,13 @@ def add_heatmap(ax, x, y, c, xinterp=None, yinterp=None, kind='linear',
             with mid and min and/or max keys. Colour format must be
             hex or rgb (array) or a named colour recognised by
             matplotlib. Default: Blues.
+        discrete : bool, optional
+            use colour bands rather than a continuously shifting colour
+            sequence. Default: False.
+        levels : int or array-like, optional
+            sets levels for discrete plots. Lists directly specify the
+            contour levels while integers specify the maximum-1 number
+            of "nice" levels to plot. Default: None.
         undercolour : str or array-like, optional
             colour for values under cmin. Default: None.
         overcolour : str or array-like, optional
@@ -113,11 +120,11 @@ def add_heatmap(ax, x, y, c, xinterp=None, yinterp=None, kind='linear',
             keyword arguments passed to matplotlib.pyplot.contour.
             Default: None
         kwargs
-            keyword arguments passed to matplotlib.pyplot.pcolormesh.
-            Defaults are defined below, which are overridden by those in
-            ``~/.config/tprc.yaml``, both of which are overridden by
-            arguments passed to this function.
-            Defaults:
+            keyword arguments passed to matplotlib.pyplot.pcolormesh or
+            matplotlib.pyplot.contourf. Defaults are defined below,
+            which are overridden by those in ``~/.config/tprc.yaml``,
+            both of which are overridden by arguments passed to this
+            function. Defaults:
 
                 rasterized: True
 
@@ -236,7 +243,21 @@ def add_heatmap(ax, x, y, c, xinterp=None, yinterp=None, kind='linear',
             y.append(y[-1] ** 2 / y[-2])
 
     # plotting
-    heat = ax.pcolormesh(x, y, c, cmap=colours, norm=cnorm, **kwargs)
+    if discrete:
+        kwargs.pop('rasterized')
+        if levels is None or levels == ():
+            heat = ax.contourf(x[:-1], y[:-1], c, cmap=colours, norm=cnorm,
+                               **kwargs)
+        else:
+            if isinstance(levels, (list, np.ndarray, tuple)) and len(levels) == 1:
+                levels = int(levels[0])
+            elif isinstance(levels, float):
+                raise TypeError('levels must be an int or array-like, not float')
+            if levels is not None and levels != ():
+                heat = ax.contourf(x[:-1], y[:-1], c, cmap=colours, norm=cnorm,
+                                   levels=levels, **kwargs)
+    else:
+        heat = ax.pcolormesh(x, y, c, cmap=colours, norm=cnorm, **kwargs)
     fig = ax.get_figure()
     if 'dos' in fig.__dict__ and fig.__dict__['dos']:
         # place colourbar outside dos
@@ -245,17 +266,39 @@ def add_heatmap(ax, x, y, c, xinterp=None, yinterp=None, kind='linear',
         cbar = plt.colorbar(heat, ax=ax, extend=extend)
 
     # contours
-    if contours is not None:
+    if contours is not None and contours != ():
         if contourkwargs == None:
             contourkwargs = {}
[…]
