### 2024-02-23 62959bf
update from JOSS review 1
simplified data retrieval in the examples
possibly fixed line endings for unix machines?
fixed broken links on the github docs. Will update the website one the main branch
updated the docstrings for the CLI options

Thanks to @epottesmith :bow:
This update is based on https://github.com/openjournals/joss-reviews/issues/6340#issuecomment-1939521565
