+++ tp/cli/options.py
@@ -3,43 +3,69 @@
 Functions
 ---------
 
-    direction_function
-    
-    doping_type_option
-    
-    doping_function
-    
-    dos_function
-    
-    input_argument
-    
-    inputs_function
-    
-    interpolate_options
-    
-    kpoints_options
-    
-    legend_options
-    
-    line_options
-    
-    fill_options
-    
-    plot_io_function
-    
-    temperature_option
-    
-    verbose_option
-
+    direction_function:
+        function for picking the --direction (-d).
+    doping_type_option:
+        function for picking the doping --type (-t).
+    doping_function:
+        function for picking the doping --concentration (-n).
+    dos_function:
+        function for setting DoS formatting.
+    inputs_function:
+        function for picking the input file(s) argument.
+    interpolate_options:
+        function for setting interpolation options.
+    kpoints_options:
+        function for handling KPOINTS files.
+    legend_options:
+        function for formatting legends.
+    line_options:
+        function for formatting line plots.
+    fill_options:
+        function for formatting fillable line plots.
+    plot_io_function:
+        function for formatting plot outputs.
+    temperature_option:
+        function for picking the --temperature (-t).
+    verbose_option:
+        function for increasing the verbosity.
     axes_limit_function
+        function for setting the axis limits.
 """
 
 import click
 
 def direction_function(multiple=False):
+    """Function to create direction options.
+        
+    Arguments
+    ---------
+
+        multiple : bool, optional
+            whether to allow multiple directions. Default: False.
+
+    Returns
+    -------
+        decorator
+            direction option decorator.
+    """
+
     default = ['avg'] if multiple else 'avg'
     def direction_option(f):
-        """Option for anisotropic data."""
+        """Option for anisotropic data.
+        
+        Options
+        -------
+
+            --direction, -d : str or array-like, optional
+                direction. Default: avg.
+
+        Returns
+        -------
+
+            decorator
+                direction option decorator.
+        """
     
         f = click.option('-d', '--direction',
                          help='Direction(s) for anisotropic data.',
@@ -56,9 +82,22 @@ def direction_function(multiple=False):
     return direction_option
 
 def doping_type_option(f):
-    """Option for doping type."""
+    """Option for doping type.
+        
+        Options
+        -------
 
-    f = click.option('-t', '--type', 'dtype',
+            --type : str, optional
+                doping type. Default: n.
+
+        Returns
+        -------
+
+            decorator
+                doping type option decorator.
+        """
+
+    f = click.option('--type', 'dtype',
                      help='Type of doping.',
                      type=click.Choice(['n', 'p']),
                      default='n',
@@ -67,9 +106,36 @@ def doping_type_option(f):
     return f
 
 def doping_function(multiple=False):
+    """Function to create doping options.
+        
+    Arguments
+    ---------
+
+        multiple : bool, optional
+            whether to allow multiple concentrations. Default: False.
+
+    Returns
+    -------
+        decorator
+            carrier concentration option decorator.
+    """
+
     default = [1e19] if multiple else 1e19
     def doping_option(f):
-        """Option for doping concentration."""
+        """Option for doping concentration.
+        
+        Options
+        -------
+
+            --concentration, -n : int or float or array-like, optional
+                carrier concentration. Default: 1e19.
+
+        Returns
+        -------
+
+            decorator
+                carrier concentration option decorator.
+        """
     
         f = click.option('-n', '--concentration', 'doping',
                          help='Doping concentration(s) (will be rounded).',
@@ -82,10 +148,53 @@ def doping_function(multiple=False):
     return doping_option
 
 def dos_function(dosargs=['-c', '--colour']):
+    """Function for creating DoS options.
+
+    Arguments
+    ---------
+
[…]
