+++ examples/04-layered-plots/tutorial-04.py
@@ -22,7 +22,7 @@ cmap = 'viridis'
 # You can ignore down to line 23!
 from os import path
 if not path.isfile(kfile) or (path.getsize(kfile) < 1024*1024*100):
-    raise Exception('File not found, please use get-data.sh')
+    raise Exception('File not found, please use get-data.sh in the folder above.')
 # Stop ignoring!
 
 # Axes
