+++ citation.cff
@@ -16,4 +16,4 @@ authors:
 title: "ThermoParser: Streamlined Analysis of Thermoelectric Properties"
 version: 3.0.0
 date-released: 2023-08-16
-url: "https://github.com/SMTG-UCL/ThermoParser"
+url: "https://github.com/smtg-bham/ThermoParser"
