### 2024-02-26 #6340 @kbspooner
Hi @espottesmith,

Thanks for your review!

 * I can't reproduce having to install matplotlib separately... Its probably to do with when I install the style sheets during the installation. If you could provide a specific error perhaps I could look into it?
 * The get_data problem must be an oversight from when I converted the examples into full tutorials. Thanks for catching it! I've now consolidated all the get_data,pys into one in the base examples directory and updated the error messages to reflect this.
 * With regards to the line endings, I've set autocrlf on in git now which should upload a version with unix line endings, but I can't test it. Could you check its worked please?
 * The broken links have been fixed in the github versions. When I push this to the main branch I'll make sure to update the website too.
 * Ahh yes I was a bit lazy with options.py 😅. It is now updated.
 * Honestly when I started this I had no idea what I was doing and found the thought of working on pymatgen intimidating. Now I'm not sure if its worth updating pymatgen as BoltzTraP is older and people can use BoltzTraP2 (which has a new python interface) or a number of higher-level codes. With respect to other codes though, while full runners might be unnecessary, perhaps expanding the setup module to write input files from calculation outputs would be useful. I haven't included this in this particular update, though.

I hope this addresses your review adequately, let me know if you have any comments/additions/etc!
