### 2024-02-26 #6340 @espottesmith
Hey @kbspooner. Appreciate the time you put into your response:

For the matplotlib issue, this is the specific error that I get (just confirmed that I'm still getting this issue with the current joss-paper branch and the main branch):

>
>               ********************************************************************************
>               Please avoid running ``setup.py`` directly.
>               Instead, use pypa/build, pypa/installer or other
>               standards-based tools.
>
>               See https://blog.ganssle.io/articles/2021/10/setup-py-deprecated.html for details.
>               ********************************************************************************
>
>       !!
>         self.initialize_options()
>       Traceback (most recent call last):
>         File "<string>", line 2, in <module>
>         File "<pip-setuptools-caller>", line 34, in <module>
>         File "/home/ewcss/software/thermoparser/setup.py", line 41, in <module>
>           setuptools.setup(
>         File "/home/ewcss/.conda/envs/thermoparser/lib/python3.11/site-packages/setuptools/__init__.py", line 103, in setup
>           return distutils.core.setup(**attrs)
>                  ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
>         File "/home/ewcss/.conda/envs/thermoparser/lib/python3.11/site-packages/setuptools/_distutils/core.py", line 185, in setup
>           return run_commands(dist)
>                  ^^^^^^^^^^^^^^^^^^
>         File "/home/ewcss/.conda/envs/thermoparser/lib/python3.11/site-packages/setuptools/_distutils/core.py", line 201, in run_commands
>           dist.run_commands()
>         File "/home/ewcss/.conda/envs/thermoparser/lib/python3.11/site-packages/setuptools/_distutils/dist.py", line 969, in run_commands
>           self.run_command(cmd)
>         File "/home/ewcss/.conda/envs/thermoparser/lib/python3.11/site-packages/setuptools/dist.py", line 989, in run_command
>           super().run_command(command)
>         File "/home/ewcss/.conda/envs/thermoparser/lib/python3.11/site-packages/setuptools/_distutils/dist.py", line 988, in run_command
>           cmd_obj.run()
>         File "/home/ewcss/.conda/envs/thermoparser/lib/python3.11/site-packages/wheel/bdist_wheel.py", line 366, in run
>           install = self.reinitialize_command("install", reinit_subcommands=True)
>                     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
>         File "/home/ewcss/.conda/envs/thermoparser/lib/python3.11/site-packages/setuptools/__init__.py", line 216, in reinitialize_command
>           cmd = _Command.reinitialize_command(self, command, reinit_subcommands)
>                 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
>         File "/home/ewcss/.conda/envs/thermoparser/lib/python3.11/site-packages/setuptools/_distutils/cmd.py", line 311, in reinitialize_command
>           return self.distribution.reinitialize_command(command, reinit_subcommands)
>                  ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
>         File "/home/ewcss/.conda/envs/thermoparser/lib/python3.11/site-packages/setuptools/_distutils/dist.py", line 941, in reinitialize_command
>           command = self.get_command_obj(command_name)
>                     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
>         File "/home/ewcss/.conda/envs/thermoparser/lib/python3.11/site-packages/setuptools/_distutils/dist.py", line 861, in get_command_obj
>           cmd_obj = self.command_obj[command] = klass(self)
>                                                 ^^^^^^^^^^^
>         File "/home/ewcss/software/thermoparser/setup.py", line 39, in __init__
>           install_style()
>         File "/home/ewcss/software/thermoparser/setup.py", line 26, in install_style
>           from matplotlib import get_configdir
>       ModuleNotFoundError: No module named 'matplotlib'
>       [end of output]
>   note: This error originates from a subprocess, and is likely not a problem with pip.
>   ERROR: Failed building wheel for tp
>   Running setup.py clean for tp
> Failed to build tp
> ERROR: Could not build wheels for tp, which is required to install pyproject.toml-based projects

In my mind, this really isn't a huge problem - the error message is very clear, and an average Python user should know how to fix this with pip/conda easily. But if you could fix it, great!

I can confirm that your changes to the line endings in your Bash script worked!

Thank you for clarifying re: pymatgen and BoltzTraP. I'm sure the pymatgen maintainers would appreciate improvements to their BoltzTraP code, but that's definitely up to you on if it's worth it or not. I agree that being able to write input files for the various codes that thermoparser interacts with would be useful, but I definitely will not ask for it, as the main purpose of this code remains post-processing, rather than managing calculation workflows.

I'm happy with the changes that you've made (assuming that you also address the other reviewers' critiques). Fixing the installation error would be ideal, but I think this code/paper could be published even with that minor error.
