### 2024-03-05 ccc671f
Updated from JOSS review 2 (mostly)
Added context to the examples
Corrected some bugs in the examples
Updated ztdiff example
updated \mathregular strings to agree with modern python
expanded tests
stopped load tests reacting to tprc.yaml
improved docs typography
miscellaneous bug fixes
paper refining

Thanks to @enricgrau
This update is mostly based on https://github.com/SMTG-Bham/ThermoParser/issues/71 and relevant to https://github.com/openjournals/joss-reviews/issues/6340
