+++ tests/test_calculate.py
@@ -1,7 +1,7 @@
 """Tests the tp.calculate module."""
 
-from re import I
 import unittest
+import numpy as np
 from tp import calculate
 
 class CumulateTest(unittest.TestCase):
@@ -22,6 +22,13 @@ class CumulateTest(unittest.TestCase):
     def test_cumulating(self):
         self.assertTrue((self.y == self.y2).all())
 
+class GaussianTest(unittest.TestCase):
+    def test_gaussian(self):
+        x = [0]
+        y = 1.0000 # to 4 dp
+        y2 = calculate.gaussian(x, x0=0, sigma=1)
+        self.assertEqual(round(y2[0], 4), y)
+
 class LorentzianTest(unittest.TestCase):
     def test_lorentzian(self):
         x = [0]
@@ -29,12 +36,46 @@ class LorentzianTest(unittest.TestCase):
         y2 = calculate.lorentzian(x, x0=0, fwhm=1)
         self.assertEqual(round(y2[0], 4), y)
 
+class LifetimeTest(unittest.TestCase):
+    def test_lifetime(self):
+        g = [[[0.25e-12 * np.pi**-1]]]
+        t = 1.0000 # to 4 dp
+        t2 = calculate.lifetime(g, use_tprc=False)
+        self.assertEqual(round(t2[0][0][0], 4), t)
+
+class MFPTest(unittest.TestCase):
+    def test_mfp(self):
+        g = [[[0.25e-12 * np.pi**-1]]]
+        v = [[[1]]]
+        l = 1.0000 # to 4 dp
+        l2 = calculate.mfp(g, v, use_tprc=False)
+        self.assertEqual(round(l2[0][0][0][0], 4), l)
+
+class FDOccupationTest(unittest.TestCase):
+    def test_fd_occupation(self):
+        e = [0]
+        ef = [0]
+        t = [300]
+        o = 0.5000
+        o2 = calculate.fd_occupation(e, t, ef, use_tprc=False)
+        self.assertEqual(round(o2[0][0][0][0], 4), o)
+
 class BEOccupationTest(unittest.TestCase):
     def test_be_occupation(self):
         f = [20]
+        t = [300]
         o = 1.5061
-        o2 = calculate.be_occupation(f, temperature=300)
-        self.assertEqual(round(o2[0], 4), o)
+        o2 = calculate.be_occupation(f, t, use_tprc=False)
+        self.assertEqual(round(float(o2[0]), 4), o)
+
+class DFDDETest(unittest.TestCase):
+    def test_dfdde(self):
+        e = [0]
+        ef = [0]
+        t = [300]
+        o = -9.6704
+        df = calculate.dfdde(e, ef, t, use_tprc=False)
+        self.assertEqual(round(df[0][0][0][0], 4), o)
 
 class PowerFactorZTTest(unittest.TestCase):
     @classmethod
@@ -43,6 +84,7 @@ class PowerFactorZTTest(unittest.TestCase):
         cls.s =   [3e6]
         cls.ke = [3.]
         cls.kl = [6.]
+        cls.k = [9.]
         cls.t =   [2.]
         cls.pf = [18.]
         cls.zt = [4.]
@@ -54,29 +96,34 @@ class PowerFactorZTTest(unittest.TestCase):
                  'meta':                            {'units':      {},
                                                      'dimensions': {}}}
 
+    def test_thermal_conductivity(self):
+        k = calculate.thermal_conductivity(self.ke, self.kl, use_tprc=False)
+        self.assertTrue((k == self.k).all())
+
     def test_power_factor(self):
-        pf2 = calculate.power_factor(conductivity=self.c, seebeck=self.s)
[…]
