### 2024-03-07 SMTG-Bham/ThermoParser#71 @kbspooner
Hi @enricgrau,

Thanks for the review!

### Manuscript

* I have expanded this sentence to give a more rigorous explanation of how it can help and how it might have advantages over similar technologies. It now reads "Thermoelectric materials, which convert heat into electricity, could be an important renewable energy source to help slow the encroaching climate crisis, not only by displacing fossil fuels, but by recycling waste heat, which makes up around 50 % of generated energy [@Firth2019]." While I don't give specific examples, I hope using a more concrete statistic makes this more suitable, otherwise like you say it may be better to leave it out entirely from what is ultimately a software paper.
* I've now added reference to the former name, as you spotted, and while those papers did use a pre-release version, as there is no evidence of this it seems fair to remove them.
* There are two libraries, BoltzTraP and BoltzTraP2, which I've now separated for clarity.
* Yeah the last sentence is pretty poor, I've rephrased it "In the future, ThermoParser could be expanded to include an increased number of analysis types and supported codes. On top of this, support for uploading experimental data into the ThermoParser format, including the appropriate metadata, could allow easier comparison of theoretical and experimental results.
* The acknowledgements are now largely converted to CRediT.

### Documentation

#### Gallery

* The links are now fixed.
* The reference has now been formatted.

#### Tutorials

* This is a (particularly) good point, I have now added introductory paragraphs to each of the tutorials that produce plots, as well as a description of the research behind the data provided in the main examples page.
* I forgot to rebuild the docs after updating them - the links should be fixed now.
* The kappa files are very large and are instead located in a Zenodo repository. As noted in the other review the instructions on accessing these weren't very clear, so the download script is now located in the base examples directory, where the files will be made accessible for all examples and tutorials.
* I cannot reproduce `raise ValueError(f"{v!r} is not a valid value for {k}") ValueError: "'#f0901f'" is not a valid value for color`, but perhaps I've fixed it in the mean time. Could you try again please?
* The formatting is now in the right language.
* While it is true that much of the code is dedicated to plotting figures for a human to analyse, I think certain parts move closer to analysing, although it depends on where you draw the line. The package ranges from pure plotting (e.g. phonon dispersions), to basic/ routine calculations (ZT calculations and plots), to more bespoke/ involved calculations which approach analysis. For example, [the avg-rates plot](https://smtg-bham.github.io/ThermoParser/avg-rates.html) uses knowldege of the Boltzmann transport equation to collapse a large amount of data ((q-mesh_x x q-mesh_y x q-mesh_z x n_bands x s_mechanisms x temperatures) points) into a handful of lines by weighting by the derivative of the Fermi-Dirac distribution wrt energy, which does much of the analysis legwork. While less complicated, [the kappa-target plot](https://smtg-bham.github.io/ThermoParser/heatmaps.html) reverses the normal thermoelectrics workflow to determine if the most expensive calculations are even worth doing, Ultimately a human must interpret the data, but some of these modules make a significant contribution in my opinion.

#### Installation

 * Looks like these and several previous error have arisen from not regenerating the docs properly after updates. I have now fixed them.

#### Python Package

* This seems to occur due to using multiple lines in the opening sentence of the docstring. I have replaced two instances of this problem in the setup/vasp.py and one in calculate.py.

### Repository

* I haven't been diligent incrementing the version with each update, so the current version is, to a large extent, guesswork. I'll try to do better in the future but there is little I can do retroactively.

### Tests

* I realise this wasn't a criticism, but I took the opportunity to make the tests more comprehensive anyhow.

I hope this addresses your problems and concerns, of course if you spot anything else or have more comments, I'll be happy to address them!

For the record, this is in relation to [this](https://github.com/openjournals/joss-reviews/issues/6340) JOSS review.
