### 2024-03-15 SMTG-Bham/ThermoParser#80 @fnattino
[joss review] Dependencies for building the documentation
In order to build the documentation, the package `sphinx_rtd_theme` is required in addition to sphinx, but not mentioned anywhere.

Also, sphinx is listed as part of the `requirements.txt` file, but you might consider creating instead a dependency set dedicated to building the documentation. See e.g. the [setuptools `extras_require` field](https://setuptools.pypa.io/en/latest/userguide/dependency_management.html#optional-dependencies).

For instance, you could add to `setup.py`:

```python
setup(
  ...
  extras_require={
    "docs"=["sphinx", "sphinx_rtd_theme"]
  }
  ...
)
```

to be able to install docs-related dependencies using:  `pip install .[docs]`

[JOSS review issue](https://github.com/openjournals/joss-reviews/issues/6340)
