### 2024-03-15 SMTG-Bham/ThermoParser#81 @fnattino
[joss review] CI
I see a `.travis.yml` file listing an old Python version in the root of the repository and a Travis CI badge with a broken link in the README - is Travis still used to automatically run tests? Maybe switch to GitHub actions (which is available for free)?

[JOSS review issue](https://github.com/openjournals/joss-reviews/issues/6340)
