### 2024-03-15 SMTG-Bham/ThermoParser#82 @fnattino
[joss review] Python API documentation
The [Python Package](https://smtg-bham.github.io/ThermoParser/modules.html) section of the documentation is quite difficult to read. Could it be improved by improving structuring and/or formatting?

For instance, functions seems to be listed twice in each module, the first time of which without any description/information (see e.g. https://smtg-bham.github.io/ThermoParser/tp.axes.html#functions). Also, some descriptions are quite "uninformative" (e.g. the top one: https://smtg-bham.github.io/ThermoParser/tp.html#tp-package), maybe these could be removed.

Also, I would name this section "Python API" in order to make it clearer what this section is actually about.

[JOSS review issue](https://github.com/openjournals/joss-reviews/issues/6340)
