### 2024-03-15 SMTG-Bham/ThermoParser#83 @fnattino
[joss review] JOSS manuscript
Just a couple of minor comments on the manuscript:

* The link to the gallery is broken (line 77), it points to an old repository?
* In the `ThermoParser` section, you mention that the code lodes data into a common format with a meta directory, that includes metadata. But what is this general format and where is this directory created?

[JOSS review issue](https://github.com/openjournals/joss-reviews/issues/6340)
