### 2024-03-15 SMTG-Bham/ThermoParser#84 @fnattino
[joss review] how is the documentation built
As far as I understand the docs are published on GitHub pages without using Jekyll (I also see the `.nojekyll` file), but there is a Jekyll config file in `docs/_config.yml`?

Also, why the folder `docs/src/_build` is checked in the repository? The build documentation is all available in `docs` , correct?

[JOSS review issue](https://github.com/openjournals/joss-reviews/issues/6340)
