### 2024-03-15 SMTG-Bham/ThermoParser#85 @fnattino
[joss review] Gallery
In the gallery I have found few examples that do not run as they are provided. Some examples:
* For [heatmaps](https://smtg-bham.github.io/ThermoParser/heatmaps.html), but also others, the `kappa-mxxxx.hdf5` is not provided, and only when checking out the Python script in the tutorial I realised how to get these files.
* Some of the commands in the [gallery main page](https://smtg-bham.github.io/ThermoParser/gallery.html#) do not follow the correct syntax required by click to provide multiple values, e.g. :  `tp plot phonons band.yaml -c '#ff8000' -d projected_dos.dat --doscolour '#ffff00' '#00ffff' ` (should instead ends with: `--doscolour '#ffff00' --doscolour  '#00ffff'`)
* Also, when fixing the problem above, the `projected_dos.dat` seems to include contributions for imaginary frequencies:
![tp-phonons](https://github.com/SMTG-Bham/ThermoParser/assets/49899980/539b4dd0-9d99-4909-a9b8-221b50d9a5ca)
* The command to run the ZTdiff (last CLI command on [this page](https://smtg-bham.github.io/ThermoParser/heatmaps.html))seems to be missing some files? `tp plot ztdiff ../data/zno/transport_75x75x75.json  ../data/zno/boltztrap.hdf5 -k ../data/zno/kappa-m404021.hdf5 ../data/zno/kappa-m404021.hdf5 -l MRTA -l CRTA
`

[JOSS review issue](https://github.com/openjournals/joss-reviews/issues/6340)
