### 2024-03-15 SMTG-Bham/ThermoParser#87 @fnattino
[joss review] Tutorial 4
In [Tutorial-04](https://smtg-bham.github.io/ThermoParser/tutorial-04.html), what does the color of the dots in the waterfall plot represent? Is it contribution to the thermal conductivity? If this is the case, should one add a colorbar as in the similar [gallery plot](https://smtg-bham.github.io/ThermoParser/waterfalls.html)? I have tried to do so using `tp.axes.small.two_h_colourbars()`, but it produced an identical plot (without colorbar).

[JOSS review issue](https://github.com/openjournals/joss-reviews/issues/6340)
