### 2024-03-15 SMTG-Bham/ThermoParser#88 @fnattino
[joss review] Tutorial 5
In [Tutorial-05](https://smtg-bham.github.io/ThermoParser/tutorial-05.html) the CLI command does not work: `-q lattice_thermal_conductivity`  and `--location 3` seem not to be available options.

[JOSS review issue](https://github.com/openjournals/joss-reviews/issues/6340)
