### 2024-03-15 SMTG-Bham/ThermoParser#89 @fnattino
[joss review] Tutorial X
While the title of the tutorial mentions VASP, the content seems to be on rather diverse topics and not necessarily related to VASP. Can a more suitable section be identified for such topics?

[JOSS review issue](https://github.com/openjournals/joss-reviews/issues/6340)
