### 2024-03-15 #6340 @fnattino
Apologies for the delay with the review @mbarzegary and @kbspooner , I have finally completed it.

Nice work @kbspooner, looks like a very useful tool! I have open individual issues with my comments/specific issues encountered,  so they can be followed up with independently.
