+++ paper.md
@@ -1,0 +1,124 @@
+---
+title: 'ThermoParser: Streamlined Analysis of Thermoelectric Properties'
+tags:
+  - Python
+  - Matplotlib
+  - thermoelectrics
+  - electronic properties
+  - transport properties
+  - materials science
+  - chemistry
+  - physics
+  - science
+  - data analysis
+  - plotting
+authors:
+  - name: Kieran B. Spooner
+    orchid: 0000-0002-4078-3148
+    affiliation: "1, 2, 3"
+  - name: Maud Einhorn
+    orchid: 0000-0003-0371-0262
+    affiliation: "2, 3"
+  - name: Daniel W. Davies
+    orchid: 0000-0003-4094-5992
+    affiliation: "2, 3"
+  - name: David O. Scanlon
+    orchid: 0000-0001-9174-8601
+    affiliation: "1, 2, 3"
+affiliations:
+ - name: School of Chemistry, University of Birmingham, Birmingham B15 2TT, United Kingdom
+   index: 1
+ - name: Department of Chemistry, University College London, 20 Gordon Street, London WC1H 0AJ, United Kingdom
+   index: 2
+ - name: Thomas Young Centre, University College London, Gower Street, London WC1E 6BT, United Kingdom
+   index: 3
+date: 28 February 2024
+bibliography: ref.bib
+---
+
+# Summary
+
+Thermoelectric materials, which convert heat into electricity, could be an important renewable energy source to help slow the encroaching climate crisis, not only by displacing fossil fuels, but by recycling waste heat, which makes up around 50 % of generated energy [@Firth2019].
+With the growing computational capacity and development of several codes to calculate the key properties of thermoelectrics, they have become an increasingly popular area of computational materials research in recent years.
+Thermal transport packages include Phonopy [@Togo2015a], Phono3py [@Togo2015], ShengBTE/ almaBTE [@Li2014], ALAMODE [@Tadano2014], TDEP [@Hellman2011] and HiPhive [@Eriksson2019]; and electronic transport packages include BoltzTraP [@Madsen2006], BoltzTraP2 [@Madsen2018], EPW [@Noffsinger2010], EPA [@Samsonidze2018], EPIC STAR [@Deng2020], AMSET [@Ganose2021], Perturbo [@Zhou2021], TOSSPB [@Poehls2022] and ElecTra [@Graziosi2023].
+While separate packages are required for such different calculations, this makes data analysis complex, needing to load in different file formats, account for different data arrangements (e.g. array shapes), and convert to consistent units, even before one begins analysing anything.
+ThermoParser deals with these time-consuming and error-prone problems by loading data from multiple codes into a consistent data format with informative metadata, and facilitates the post-processing of thermoelectric properties by using this to accurately calculate and visualise them through an easy-to-use command-line interface (CLI) and a fully customisable Python package.
+Some of its utility can be seen by its use in the literature (sometimes under its former name, ThermoPlotter) [@Kavanagh2021;@Spooner2021;@Brlec2022;@HerringRodriguez2023;@Willis2023;@Han2024].
+
+# Statement of Need
+
+To the best of our knowledge no package exists for processing data from the vastly different sets of calculations needed to study thermoelectrics computationally.
+While there are several codes to aid with individual aspects of thermoelectric calculations, such as the inbuilt analysis sections of Phonopy and AMSET or packages such as Phono3py-Power-Tools [@Skelton2020] and sumo [@Ganose2018], they are specialised to either the phononic or electronic side.
+ThermoParser brings three key novelties to the existing software landscape:
+
+* The automatic parsing of outputs from multiple codes for both electronic and phononic calculations;
+* A data system which is transparent in the origin, arrangement and units of the data, customisable, and accessible regardless of Python aptitude;
[…]
