+++ ref.bib
@@ -1,0 +1,355 @@
+% Encoding: UTF-8
+
+@Article{Hellman2011,
+  author    = {Hellman, Olle and Abrikosov, I A and Simak, SI},
+  journal   = {Physical Review B},
+  title     = {Lattice Dynamics of Anharmonic Solids from First Principles},
+  year      = {2011},
+  number    = {18},
+  pages     = {180301},
+  volume    = {84},
+  doi       = {10.1103/PhysRevB.84.180301},
+  publisher = {APS},
+}
+
+@Article{Hellman2013,
+  author    = {Hellman, Olle and Abrikosov, Igor A},
+  journal   = {Physical Review B},
+  title     = {Temperature-Dependent Effective Third-Order Interatomic Force Constants From First Principles},
+  year      = {2013},
+  number    = {14},
+  pages     = {144301},
+  volume    = {88},
+  doi       = {10.1103/PhysRevB.88.144301},
+  publisher = {APS},
+}
+
+@Article{Hellman2013a,
+  author    = {Hellman, Olle and Steneteg, Peter and Abrikosov, Igor A and Simak, Sergei I},
+  journal   = {Physical Review B},
+  title     = {Temperature Dependent Effective Potential Method for Accurate Free Energy Calculations of Solids},
+  year      = {2013},
+  number    = {10},
+  pages     = {104111},
+  volume    = {87},
+  doi       = {10.1103/PhysRevB.87.104111},
+  publisher = {APS},
+}
+
+@Article{Togo2015,
+  author    = {Togo, Atsushi and Chaput, Laurent and Tanaka, Isao},
+  journal   = {Physical Review B},
+  title     = {Distributions of Phonon Lifetimes in Brillouin Zones},
+  year      = {2015},
+  month     = {Mar},
+  number    = {9},
+  pages     = {094306},
+  volume    = {91},
+  doi       = {10.1103/PhysRevB.91.094306},
+  issue     = {9},
+  numpages  = {31},
+  publisher = {American Physical Society},
+}
+
+@Article{Togo2015a,
+  author    = {Togo, Atsushi and Tanaka, Isao},
+  journal   = {Scripta Materialia},
+  title     = {First Principles Phonon Calculations in Materials Science},
+  year      = {2015},
+  pages     = {1},
+  volume    = {108},
+  doi       = {10.1016/j.scriptamat.2015.07.021},
+  publisher = {Elsevier},
+}
+
+@Article{Li2014,
+  author  = {Wu Li and Jes\'us Carrete and Nebil A. Katcho and Natalio Mingo},
+  journal = {Computer Physics Communications},
+  title   = {{ShengBTE:} a Solver of the {B}oltzmann Transport Equation for Phonons},
+  year    = {2014},
+  number  = {6},
+  pages   = {1747},
+  volume  = {185},
+  doi     = {10.1016/j.cpc.2014.02.015},
+}
+
+@Article{Einhorn2020,
+  author    = {Maud Einhorn and Benjamin A. D. Williamson and David O. Scanlon},
+  journal   = {Journal of Materials Chemistry A},
+  title     = {Computational Prediction of the Thermoelectric Performance of {LaZnOPn} ({Pn} = {P}, {As})},
+  year      = {2020},
+  number    = {16},
+  pages     = {7914},
+  volume    = {8},
+  doi       = {10.1039/d0ta00690d},
+  publisher = {Royal Society of Chemistry ({RSC})},
+}
+
+@Article{Spooner2020,
+  author    = {Kieran B. Spooner and Alex M. Ganose and David O. Scanlon},
+  journal   = {Journal of Materials Chemistry A},
[…]
