### 2024-03-18 8dfd55b
updated setup
added missing dependencies
made the requirements for the docs and tests extras installed only on request
updated the docs about docs and tests

in response to @fnattino's JOSS review in this issue: https://github.com/SMTG-Bham/ThermoParser/issues/80
