+++ requirements.txt
@@ -5,4 +5,3 @@ numpy >=1.14.3
 pymatgen >=2020.1.28
 pyyaml
 scipy
-sphinx
