+++ README.rst
@@ -154,7 +154,9 @@ while contributions can be made using the `fork and pull`_ approach.
 Contributions should include comprehensive docstrings, and where
 appropriate `examples`_, further `documentation`_ and `tests`_ are greatly
 appreciated. Documentation uses the `sphinx`_ package, and can be built from
-the ``docs`` directory with ``sphinx-build -b html src/ .``.
+the ``docs`` directory with ``sphinx-build -b html src/ .``. In order to build
+the docs, download the extra dependencies with, e.g., ``pip install .[docs]``
+from the ThermoParser directory.
 
 .. _issue tracker: https://github.com/smtg-bham/ThermoParser/issues
 .. _fork and pull: https://guides.github.com/activities/forking
@@ -165,7 +167,9 @@ Testing
 -------
 
 Tests use the `unittest`_ package, and can be run from the test directory
-with ``python3 -m unittest``.
+with ``python3 -m unittest``. If you don't already have unittest installed, it
+can be directly with ``pip install unittest`` or, e.g., ``pip install .[tests]``
+from the ThermoParser directory.
 
 .. _unittest: https://docs.python.org/3/library/unittest.html
 
