### 2024-03-18 SMTG-Bham/ThermoParser#80 @kbspooner
Ahh this is handy, I've done this for the docs and the tests and updated the relevent docs sections, currently in the bug-fix branch.
