### 2024-03-18 3c6c19c
Create python-app.yml
use github actions for tests instead of travis
