### 2024-03-18 2cd20e4
Delete .travis.yml
switch to github actions
