+++ .travis.yml
@@ -1,15 +1,0 @@
-language: python
-
-python:
-  - "3.6"
-
-install:
-  - pip install --upgrade pip
-  -  python -m pip install -r requirements.txt .
-  
-branches:
-  only:
-  - /.*/
-
-script: 
-  - cd tests/ && python -m unittest
