### 2024-03-18 5873bed
Update python-app.yml
typo
