### 2024-03-18 e1f6059
update github actions
added build and docs workflows
added status badges
