+++ .github/workflows/docs.yml
@@ -1,0 +1,32 @@
+# Build ThermoParser's docs
+
+name: docs
+
+on:
+  push:
+    branches: [ "master" ]
+  pull_request:
+    branches: [ "master" ]
+
+permissions:
+  contents: read
+
+jobs:
+  build:
+
+    runs-on: ubuntu-latest
+
+    steps:
+    - uses: actions/checkout@v3
+    - name: Set up Python 3.10
+      uses: actions/setup-python@v3
+      with:
+        python-version: "3.10"
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install sphinx sphinx_rtd_theme
+        if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+    - name: Build docs with sphinx
+      run: |
+        cd docs; sphinx-build -b html src/ .
