+++ .github/workflows/docs.yml
@@ -5,8 +5,6 @@ name: docs
 on:
   push:
     branches: [ "master" ]
-  pull_request:
-    branches: [ "master" ]
 
 permissions:
   contents: read
@@ -25,8 +23,12 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install sphinx sphinx_rtd_theme
-        if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+        pip install -e '.[docs]'
     - name: Build docs with sphinx
       run: |
         cd docs; sphinx-build -b html src/ .
+    - name: Deploy
+      uses: peaceiris/actions-gh-pages@v3
+      with:
+        github_token: ${{ secrets.GITHUB_TOKEN }}
+        publish_dir: docs/src/
