### 2024-03-18 432da29
decouple docs from build
also make installation consistent across workflows
