+++ .github/workflows/tests.yml
@@ -25,8 +25,8 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install unitest
-        if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+        python -m pip install unitest
+        if [ -f requirements.txt ]; then python -m pip install -r requirements.txt; fi
     - name: Test with unittest
       run: |
         python3 -m unittest
