+++ .github/workflows/docs.yml
@@ -23,7 +23,7 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install -e '.[docs]'
+        python -m pip install sphinx sphinx_rtd_theme
     - name: Build docs with sphinx
       run: |
         cd docs; sphinx-build -b html src/ .
