+++ .github/workflows/docs.yml
@@ -8,6 +8,7 @@ on:
 
 permissions:
   contents: read
+  pages: write
 
 jobs:
   build:
