+++ .github/workflows/docs.yml
@@ -31,5 +31,5 @@ jobs:
     - name: Deploy
       uses: peaceiris/actions-gh-pages@v3
       with:
-        github_token: ${{ secrets.GITHUB_TOKEN }}
-        publish_dir: docs/src/
+        github_token: ${{ secrets.ACTIONS_DEPLOY_KEY }}
+        publish_dir: docs/
