### 2024-03-18 SMTG-Bham/ThermoParser#81 @kbspooner
Didn't reallise it had expired, switched to github actions. I also added actions for building and updating docs, and added badges for all three on the front page. It also recreated the install error @espottesmith encountered but I couldn't previously recreate, which I'll try to fix later.
