### 2024-03-18 0ea5463
update docs
remove duplication of sections of the docs, minor editing
