+++ tp/data/run.py
@@ -1,10 +1,10 @@
-"""Code running tools.
+"""External code running tools."""
 
-Functions
----------
-
-    boltztrap
-"""
+#Functions
+#---------
+#
+#    boltztrap
+#"""
 
 import numpy as np
 import shutil
