+++ tp/setup/__init__.py
@@ -1,10 +1,10 @@
-"""Tools for setting up calculations and inputs
+"""Tools for setting up calculations and inputs."""
 
-Modules
--------
-
-    vasp
-        currently contains KPOINTS generators
-"""
+#Modules
+#-------
+#
+#    vasp
+#        currently contains KPOINTS generators
+#"""
 
 from . import vasp
