+++ tp/cli/__init__.py
@@ -1,12 +1,12 @@
-"""Provides CLI.
+"""Provides a command-line interface."""
 
-Modules
--------
-
-    cli
-        Command line interface functions
-    options
-        Option groups for cli functions.
-"""
+#Modules
+#-------
+#
+#    cli
+#        Command line interface functions
+#    options
+#        Option groups for cli functions.
+#"""
 
 from . import cli, options
