### 2024-03-18 SMTG-Bham/ThermoParser#82 @kbspooner
Ah this seems to be a hangover from before using sphinx. I've removed all of the duplication (I think) and modified some of the docstrings, as well as the Python API title.
