### 2024-03-18 SMTG-Bham/ThermoParser#83 @kbspooner
I have corrected the former, and the latter is mistyped, it should've said dictionary, which is a subdictionary in any data you load through tp, and this has now also been corrected.
