### 2024-03-19 SMTG-Bham/ThermoParser#71 @enricgrau
Thank you for your comprehensive response. I’m quite satisfied with the changes you’ve made so far. I like the current state of the documentation and the manuscript. Here are some additional minor observations.

**Installation**
This time I came across some minor issue with the commeand ``pip install .`` with this error:

```
ModuleNotFoundError: No module named 'matplotlib'
      [end of output]

  ERROR: Failed building wheel for tp
Failed to build tp
ERROR: Could not build wheels for tp, which is required to install pyproject.toml-based projects
```

I fixed it with the following command before retrying: ``pip install setuptools wheel``

I’m now in a different machine, so I’m not sure it’s a local issue but I tend to believe this. Because of this, and as it is an easy fix, I don’t think is a necessary fix for being accepted. You may have a quick look into this.

**Data**
I was now able to download the data. However, I was only able to do so because you mentioned it here. I did not find any explicit mention on where or how to downloading the data in the documentation. I strongly recommend you include just a small instruction to do this (either manually or/and with the included in script). I might have missed this however, but it just was not clear to me.

**Tutorials**
I only had problems with tutorial 2. The script keeps running infinitely and I get the following error every couple of seconds (this I was able to reproduce in both of my machines):
```
  File "C:\Users\User\AppData\Local\Programs\Python\Python311\Lib\multiprocessing\spawn.py", line 140, in _check_not_importing_main
    raise RuntimeError('''
RuntimeError:
        An attempt has been made to start a new process before the
        current process has finished its bootstrapping phase.

        This probably means that you are not using fork to start your
        child processes and you have forgotten to use the proper idiom
        in the main module:

            if __name__ == '__main__':
                freeze_support()
                ...

        The "freeze_support()" line can be omitted if the program
        is not going to be frozen to produce an executable.

        To fix this issue, refer to the "Safe importing of main module"
        section in https://docs.python.org/3/library/multiprocessing.html
```

**Repository**
I understand there is nothing you can do retroactively. However, I insist that it would be great to keep this consistent in the future. I see now that the latest uploaded version is ``3.1.2``, which is great. I encourage you to make a new release matching the latest real version once the JOSS review is done. This so it incorporates all the changes and it accurately reflects your work and progress.
