### 2024-03-20 b7a0e2c
Improve docs maintainability
import scripts into the docs to avoid needing to (or forgetting to) repeat update them
update locations of  files in the docs to avoid duplication
