+++ examples/phonons/plot-phonons.sh
@@ -1,0 +1,1 @@
+tp plot phonons ../data/zno/band.yaml --dos ../data/zno/projected_dos.dat --poscar ../data/zno/POSCAR -doscolour '#d46ef9' --doscolour '#7b8eff' -c '#f0901f'
