### 2024-03-20 ed84b02
Use correct filepath for ztdiff example
