### 2024-03-20 SMTG-Bham/ThermoParser#85 @kbspooner
* I realised I only added the explanation of this in the github readme, which didn't make it into the website docs. The relevent text has been copied into the gallery page, and in both cases a title has been added highlighting it.
* Looks like when I updated from argparse to click I updated the main tutorials and examples, but missed some of the gallery code. I've now made it so the gallery imports the bash code directly from the examples, so this won't happen again. I also updated other parts of the docs similarly.
* I actually noticed this the other day and the offending file has been replaced with the converged publication-quality data.
* Whoops, that was meant to be basno3 not zno. Now corrected.

These updates are currently in the bug-fix branch.
