### 2024-03-20 203db1b
examples mac bug fix
fixed a bug in the examples where scripts would crash if multiprocessing on mac
