+++ examples/projected-phonons/plot-wideband.py
@@ -12,17 +12,24 @@ if not path.isfile(kappafile) or (path.getsize(kappafile) < 1024*1024*100):
 
 colour = ['#000000', '#ff0000']
 
-# Axes
-fig, ax, add_legend = tp.axes.large.one('dark_background')
+# running this section as a function is particularly important for mac users
+# due to its use of multiprocessing (this is the case for all projected phonon
+# plots including alt_phonons and wideband)
+def main():
+    # Axes
+    fig, ax, add_legend = tp.axes.large.one('dark_background')
+    
+    # Load
+    data = tp.data.load.phono3py(kappafile, quantities='wideband')
+    pdata = tp.data.load.phonopy_dispersion(phile)
+    
+    # Add
+    tp.plot.phonons.add_wideband(ax, data, pdata, temperature=temperature,
+                                 colour=colour, poscar=poscar)
+    
+    # Save
+    fig.savefig('wideband.pdf')
+    fig.savefig('wideband.png')
 
-# Load
-data = tp.data.load.phono3py(kappafile, quantities='wideband')
-pdata = tp.data.load.phonopy_dispersion(phile)
-
-# Add
-tp.plot.phonons.add_wideband(ax, data, pdata, temperature=temperature,
-                             colour=colour, poscar=poscar)
-
-# Save
-fig.savefig('wideband.pdf')
-fig.savefig('wideband.png')
+if __name__ == "__main__":
+    main()
