### 2024-03-20 SMTG-Bham/ThermoParser#86 @kbspooner
Thanks for the fix! That will probably occur on all the projected phonons functions, so I've added it to those examples too.
