### 2024-03-20 SMTG-Bham/ThermoParser#87 @kbspooner
In this case it just the reflects the band index, which is now written in the introduction to the tutorial. The ``_colourbar`` axes modules only leave space for a colourbar, only the plot functions (or the user) makes the actual colourbar, which is returned by the ``add_projected_waterfall`` function, but not by the ``add_waterfall`` function. This is the default behaviour if a colourmap is specified in the ``add_waterfall`` function, but a single colour will give a single-colour plot (this is now also included in the tutorial).
