### 2024-03-20 SMTG-Bham/ThermoParser#88 @kbspooner
It works for me, but I think I fixed this quite recently, so perhaps you have an older version. I'll consolidate all these fixes into one version soon, and then could you try again please?
