### 2024-03-20 SMTG-Bham/ThermoParser#89 @kbspooner
Ahh true. I've taken the lazy way out and renamed the whole section to miscellaneous assists.
