### 2024-03-20 SMTG-Bham/ThermoParser#84 @kbspooner
The docs folders were a mess! I've extended the changes mentioned in [issue 85](https://github.com/SMTG-Bham/ThermoParser/issues/85) to streamline the docs, and deleted some extra directories which must have been generated by me compiling them wrong and then never deleted.
