### 2024-03-20 #6340 @kbspooner
Hi @fnattino,

No worries, it was worth the wait!

I've now responded to each of your issues, hopefully to your satisfaction. I couldn't recreate [issue 88](https://github.com/SMTG-Bham/ThermoParser/issues/88), but I think I noticed and fixed it while updating based on another review, so could you check again please? All the fixes are in the master branch now. Naturally, any further comments etc. are welcome!
