### 2024-03-21 f7f199b
fix rare installation bug and docs updates
