+++ requirements.txt
@@ -5,3 +5,5 @@ numpy >=1.14.3
 pymatgen >=2020.1.28
 pyyaml
 scipy
+setuptools
+wheel
