+++ README.rst
@@ -13,7 +13,7 @@
     | |   |___   ___  |___  |_____   ___ :|___/  ___   |___   ___   ___  |___ \
     | |   |   | /   \ |   \ |  |  | /   \:|     /   |  |   \ /     /   \ |   \/
     | |   |   | |___/ |     |  |  | |   |:|     |   |  |     \___  |___/ |    \_____
-    | |   |   | \___  |     |  |  | \___/:|     \___|_ |         \ \___  |    :3.1.2\
+    | |   |   | \___  |     |  |  | \___/:|     \___|_ |         \ \___  |    :3.1.3\
     |____________________________________:_______________________/____________:_____/
 
 
