### 2024-03-27 SMTG-Bham/ThermoParser#80 @fnattino
Great, nice setup! Concerning the tests: I think `unittest` is already part of the standard library, you should not need any additional dependency to run tests via `python -m unittest`, correct? Also note the different spelling of the `unitest` package you have now as extra dependency (only one "t"), not sure whether this is actually needed?
