### 2024-03-27 SMTG-Bham/ThermoParser#82 @fnattino
Nice, much more readable now, thanks!
