### 2024-03-27 SMTG-Bham/ThermoParser#87 @fnattino
Thanks, all clear now!
