### 2024-03-27 SMTG-Bham/ThermoParser#88 @fnattino
It works now, thanks!
