### 2024-03-27 SMTG-Bham/ThermoParser#89 @fnattino
Looks good, now a user reading the docs won't be surprised when opening that section. Thanks!
