### 2024-03-27 #6340 @fnattino
Thanks for addressing the issues @kbspooner, nice work!

@mbarzegary my comments have been addressed to large extend, there are only a couple of issues still pending. I have left an additional suggestion for improvement as an issue to the package, but haven't linked it here since I consider it optional and up to the authors whether they want to consider it.
