### 2024-03-28 SMTG-Bham/ThermoParser#71 @enricgrau
No problem! Thank you for your hard work. I'll close this issue and recommend for publication.
