### 2024-03-29 4b05c8f
bug fixes
based on https://github.com/SMTG-Bham/ThermoParser/issues/85 by @fnattino
