### 2024-03-29 SMTG-Bham/ThermoParser#85 @kbspooner
Thanks for catching these! All are now fixed. Not sure why the last one works for me and not you, but the suggested change has the same effect and if it makes it more robust that must be a good thing.
