### 2024-03-29 SMTG-Bham/ThermoParser#80 @kbspooner
Ahh is that what happened. I got an error when I tried to install unittest, presumably because its built-in as you say, and it suggested unitest instead, so I blindly followed. I think I've done that before actually, but I will remember next time! Thanks!
