### 2024-04-02 ea5ba1c
version transparency and help help
version can now be shown with tp.__version__ or by passing the --version option to any tp cli function
help is now shown with -h as well as --help
docs updates
