+++ setup.py
@@ -3,10 +3,10 @@
 __name__ =       'ThermoParser'
 __author__ =     'Kieran B. Spooner'
 __copyright__ =  'Copyright Scanlon Materials Theory Group (2024)'
-__version__ =    '3.1.3'
+__version__ =    '3.1.4' # Remember to update the main README header too!
 __maintainer__ = 'Kieran B. Spooner'
 __email__ =      'k.b.spooner@bham.ac.uk'
-__date__ =       'March 21st 2024'
+__date__ =       'April 1st 2024'
 
 import os
 import setuptools
@@ -63,6 +63,6 @@ setuptools.setup(
     install_requires=['click', 'h5py', 'matplotlib', 'numpy', 'pymatgen',
                       'pyyaml', 'scipy', 'setuptools', 'wheel'],
     extras_require={'docs':  ['sphinx', 'sphinx_rtd_theme']},
-    python_requires='>=3.6',
+    python_requires='>=3.8',
     cmdclass={'install': PostInstallMoveFile},
     entry_points={'console_scripts': ['tp = tp.cli.cli:tp_cli']})
