### 2024-04-03 SMTG-Bham/ThermoParser#85 @fnattino
Thanks - I can confirm they all work now!
