### 2024-04-17 9ca5fdd
minor edits on paper.md
