### 2024-04-17 #6340 @mbarzegary
@kbspooner given the green light of the reviewers, we will now work towards processing this for acceptance in JOSS. So please
* Merge my PR with some minor edits
* Work on the author's points of the final checklist I created ☝️
* Make sure that the acronym `DoS` has a definition in the text
