### 2024-04-17 1ddc8d5
minor paper edit
defines DoS
