### 2024-04-17 #6340 @kbspooner
Thanks!
I've merged your PR, checked the paper, archived on Zenodo with the appropriate metadata.
About the missing DOI in the references, there is none for that software, do I need to do something about that?

Version:    3.1.4
DOI:         10.5281/zenodo.10987923
