### 2024-04-19 #6340 @mbarzegary
@kbspooner I see that the mentioned software doesn't have any other mean of citation or a even a version number unfortunately. Moreover, it seems that you needed to cite it since you wanted to compare you code with it, so we cannot mark it as an unnecessary citation and replace it with a link. As a result, I think it's fine to proceed with how it is now.
