### 2024-05-01 #6340 @Kevin-Mattheus-Moerman
@kbspooner as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. I have checked this review, your repository, the archive link, and the paper. Most seems in order, I'll just ask one reviewer for clarification before we move on.
